real optimizations

+++ cpp/example.cpp
@@ -12,8 +12,8 @@ extern "C" {  // avoid name mangling
     EMSCRIPTEN_KEEPALIVE
     // array argument types MUST correspond to expected array type,
     // otherwise you will never get right values due to wasm memory little-endianess
-    void *baz(long long *a, long long *b, size_t length) { 
-        long long *c = new long long[length];
+    void *baz(int *a, int *b, size_t length) { 
+        int *c = new int[length];
         for (size_t i = 0; i < length; i++) {
             c[i] = a[i] * b[i];
         }
